Rewriting terms in the BFS for conversion is now performed by multiple threads.
Each thread is assigned a subset of all equations, and rewrites all terms in the sets leftTerms and rightTerms using only those equations.

This greatly improves the speed of the program.

Current implementation is quite basic: a fixed number of threads is used, and dividing equations amongst threads might not necessarily be the best solution